Refactor global UTC API and tighten debug prints
Replaced the `Global` singleton in `global.h` with `global` namespace inline helpers for UTC offset get/set/validation, simplifying access while keeping behavior intact. Updated multiple debug/display headers to use explicit casts that match `printf` format specifiers (unsigned values and pointer casts), renamed a few local pointers for clarity, and normalized literals. Also applied minor readability cleanups in `json_parser.h` by expanding single-line guard clauses, plus a yearly copyright bump in `debug_dump.h`.

+++ common/include/firmware/debug/debug_i2cdetect.h
@@ -44,7 +44,7 @@ void Detect() {
     puts("\n     0  1  2  3  4  5  6  7  8  9  a  b  c  d  e  f");
 
     for (uint32_t i = 0; i < 128; i = (i + 16)) {
-        printf("%02x: ", i);
+        printf("%02x: ", static_cast<unsigned>(i));
         for (uint32_t j = 0; j < 16; j++) {
             // Skip unwanted addresses
             if ((i + j < kFirst) || (i + j > kLast)) {
@@ -53,7 +53,7 @@ void Detect() {
             }
 
             if (::i2c::IsConnected(static_cast<uint8_t>(i + j))) {
-                printf("%02x ", i + j);
+                printf("%02x ", static_cast<unsigned>(i + j));
             } else {
                 printf("-- ");
             }
